% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{Game-class}
\alias{Game}
\alias{Game-class}
\title{Game framework class}
\description{
The class is the base class of the gmae framework.
You need extend the class within subclass.
}
\examples{
\dontrun{
 Snake<-setRefClass("Snake",contains="Game")
}
}

